# entitlement-service/entitlement-service-openapi.yml
openapi: 3.0.3
info:
  title: Entitlement Service API
  description: |
    NextGen Workflow Entitlement Service - Central authority for authentication, authorization, and user management.
    
    This service provides:
    - Session-based authentication
    - Cerbos policy-based authorization
    - User and role management
    - Business application management
    
    **Authentication**: All endpoints except `/api/auth/login` require session-based authentication via `X-Session-Id` header.
    
    **Authorization**: Fine-grained access control is provided through Cerbos policy engine integration.
  version: 1.0.0
  contact:
    name: NextGen Workflow Team
servers:
  - url: http://localhost:8081
    description: Development server
...
paths:
...
  /api/entitlements/business-app-roles:
    get:
      tags:
        - Business App Roles
      summary: Get all business app roles
      description: Retrieve all business application roles
      responses:
        '200':
          description: Business app roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BusinessAppRole'

    post:
      tags:
        - Business App Roles
      summary: Create business app role
      description: Create a new business application role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessAppRole'
      responses:
        '201':
          description: Business app role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAppRole'
        '400':
          description: Invalid business app role data
  /api/entitlements/business-app-roles/{id}:
    get:
      tags:
        - Business App Roles
      summary: Get business app role by ID
      description: Retrieve a specific business app role by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business app role ID
      responses:
        '200':
          description: Business app role found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAppRole'
        '404':
          description: Business app role not found

    put:
      tags:
        - Business App Roles
      summary: Update business app role
      description: Update an existing business app role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business app role ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessAppRole'
      responses:
        '200':
          description: Business app role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAppRole'
        '404':
          description: Business app role not found

    delete:
      tags:
        - Business App Roles
      summary: Delete business app role
      description: Delete a business app role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business app role ID
      responses:
        '204':
          description: Business app role deleted successfully
        '404':
          description: Business app role not found
...
  /api/entitlements/business-app-roles/business-application/{businessAppId}/role/{roleName}:
    get:
      tags:
        - Business App Roles
      summary: Get role by business application and name
      description: Retrieve a specific role by business application and role name
      parameters:
        - name: businessAppId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business application ID
        - name: roleName
          in: path
          required: true
          schema:
            type: string
          description: Role name
      responses:
        '200':
          description: Role found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAppRole'
        '404':
          description: Role not found
...
  /api/entitlements/business-app-roles/{id}/activate:
    patch:
      tags:
        - Business App Roles
      summary: Activate business app role
      description: Activate a business app role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business app role ID
      responses:
        '200':
          description: Business app role activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAppRole'
        '404':
          description: Business app role not found
  /api/entitlements/business-app-roles/{id}/deactivate:
    patch:
      tags:
        - Business App Roles
      summary: Deactivate business app role
      description: Deactivate a business app role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Business app role ID
      responses:
        '200':
          description: Business app role deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessAppRole'
        '404':
          description: Business app role not found

  # User Business App Roles endpoints
...
  /api/entitlements/user-business-app-roles/role/{roleId}:
    get:
      tags:
        - User Business App Roles
      summary: Get users by role ID
      description: Retrieve all users assigned to a specific role
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Role ID
      responses:
        '200':
          description: Users with role retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBusinessAppRole'
  /api/entitlements/user-business-app-roles/role/{roleId}/active:
    get:
      tags:
        - User Business App Roles
      summary: Get active users by role ID
      description: Retrieve all active users assigned to a specific role
      parameters:
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Role ID
      responses:
        '200':
          description: Active users with role retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserBusinessAppRole'
  /api/entitlements/user-business-app-roles/user/{userId}/role/{roleId}:
    get:
      tags:
        - User Business App Roles
      summary: Get specific user role assignment
      description: Retrieve a specific user role assignment by user ID and role ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID format)
        - name: roleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Role ID
      responses:
        '200':
          description: User role assignment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBusinessAppRole'
        '404':
          description: User role assignment not found
...
components:
...
  schemas:
...
    BusinessApplication:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        businessAppName:
          type: string
          example: "onecms"
        description:
          type: string
          example: "OneCMS Case Management System"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        metadata:
          type: object
          additionalProperties: true
          example:
            version: "1.0.0"
            owner: "Case Management Team"
    BusinessAppRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        businessApplication:
          $ref: '#/components/schemas/BusinessApplication'
        roleName:
          type: string
          example: "INTAKE_ANALYST"
        roleDisplayName:
          type: string
          example: "Intake Analyst"
        description:
          type: string
          example: "Analyst responsible for case intake and initial review"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        metadata:
          type: object
          additionalProperties: true
          example:
            permissions: ["READ", "CREATE"]
            queue: "intake-analyst-queue"
    UserBusinessAppRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        businessAppRole:
          $ref: '#/components/schemas/BusinessAppRole'
        isActive:
          type: boolean
          example: true
...